Change SaveState API to use Buffers
This modifies gyVmuDeviceSaveState and gyVmuDeviceLoadState to take in a
buffer and a size rather than a file path. File IO is now left to the
user.
(Alternatively, we could add file IO functions instead.)

This also fixes an unrelated crash that came from loading a fully zeroed
save state, especially if it was from a 32-bit arch on a 64 bit machine
or vise-versa.