refactor: move repeated code snippets into a helper function
The changes are to replace the repeated code in all of the tests to a helper function that can be used instead.

NOTE: I only did it for a couple of functions so the reviewer can take a look and see if it's something they'd be willing to accept. If accepted, I will take the time to implement the helper in more functions across all the tests.